Vincula manutenção ao veículo por ID e registra quilometragem de retirada/devolução — ManutencaoCreateDTO passa a receber veiculoId em vez da entidade Veiculo, ManutencaoService valida status do veículo antes de criar manutenção, LocacaoService grava kmRetirada e kmDevolucao, DTOs de Veiculo padronizam quilometragem para Double e adiciona testes unitários dos services

+++ src/main/java/io/github/arthursilvagbs/Locacao/de/Carros/dto/veiculo/VeiculoResponseDTO.java
@@ -18,7 +18,7 @@ public record VeiculoResponseDTO(
         LocalDateTime ano,
         String cor,
         CategoriaVeiculo categoriaVeiculo,
-        Integer quilometragem,
+        Double quilometragem,
         LocalDateTime createdAt
 ) {
 }
